complete file breakout
