+++ FileTransferCS/FileTransferCS.vcxproj
@@ -141,15 +141,28 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="FileReader.cpp" />
+    <ClCompile Include="FileTransferClient.cpp" />
     <ClCompile Include="FileTransferCS.cpp" />
+    <ClCompile Include="FileTransferServer.cpp" />
     <ClCompile Include="TransactionUnit.cpp" />
+    <ClCompile Include="UDPReceiver.cpp" />
+    <ClCompile Include="UDPUnreliableSender.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="FileReader.h" />
+    <ClInclude Include="FileTransferClient.h" />
+    <ClInclude Include="FileTransferServer.h" />
     <ClInclude Include="ILogger.h" />
+    <ClInclude Include="IReader.h" />
+    <ClInclude Include="IReceiver.h" />
+    <ClInclude Include="ISender.h" />
     <ClInclude Include="IWorkerThreadPool.h" />
     <ClInclude Include="SimpleLogger.h" />
     <ClInclude Include="TransactionManager.h" />
     <ClInclude Include="TransactionUnit.h" />
+    <ClInclude Include="UDPReceiver.h" />
+    <ClInclude Include="UDPUnreliableSender.h" />
     <ClInclude Include="WorkerThreadPool.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
