+++ FileTransferCS/IReceiver.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <functional>
+#include <vector>
+
+class IReceiver
+{
+public:
+   virtual void Receive(std::function<void(std::vector<char>)> callback) = 0;
+};
