+++ FileTransferCS/ISender.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <vector>
+
+class ISender
+{
+public:
+   virtual void Send(const std::vector<char>& s) = 0;
+};
+
